png saves asynchronously
The png file is now saved asynchronously, significantly improving the processing time of each image. The first takes 250, but thereafter only 60-70ms per image. When you are finished recording, press the home button. The app will remain "open" but appear frozen as it saves on the pngs still in the queue. After all photos that were classified are also saved, the app will close. Do not press the start button until you see all of the photos are saved.